chore: move modal dir
