+++ feat/time-capsule-detail/src/main/java/com/emotionstorage/time_capsule_detail/ui/modal/SaveChangesModal.kt
@@ -5,7 +5,7 @@ import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.emotionstorage.ui.component.Modal
+import com.emotionstorage.ui.component.modal.Modal
 
 @Composable
 fun SaveChangesModal(
