+++ feat/time-capsule-detail/src/main/java/com/emotionstorage/time_capsule_detail/ui/modal/TimeCapsuleUnlockModal.kt
@@ -26,7 +26,7 @@ import com.emotionstorage.common.getDaysBetween
 import com.emotionstorage.ui.R
 import com.emotionstorage.ui.component.CountDownTimer
 import com.emotionstorage.ui.component.button.CtaButton
-import com.emotionstorage.ui.component.Modal
+import com.emotionstorage.ui.component.modal.Modal
 import com.emotionstorage.ui.theme.MooiTheme
 import java.time.LocalDate
 import java.time.LocalDateTime
